+++ TLP/src/main/java/com/stenaeke/TLP/dtos/subcategory/UpdateSubcategoryCourse.java
@@ -1,11 +1,9 @@
 package com.stenaeke.TLP.dtos.subcategory;
 
-import com.stenaeke.TLP.domain.Course;
-import com.stenaeke.TLP.domain.Subcategory;
 import lombok.Data;
 
 @Data
 public class UpdateSubcategoryCourse {
-    private int courseId;
+    private Long courseId;
 
 }
