+++ TLP/src/main/java/com/stenaeke/TLP/repositories/ModuleRepository.java
@@ -3,5 +3,5 @@ package com.stenaeke.TLP.repositories;
 import com.stenaeke.TLP.domain.Module;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ModuleRepository extends JpaRepository<Module, Integer> {
+public interface ModuleRepository extends JpaRepository<Module, Long> {
 }
