+++ TLP/src/main/java/com/stenaeke/TLP/repositories/SubcategoryRepository.java
@@ -3,5 +3,5 @@ package com.stenaeke.TLP.repositories;
 import com.stenaeke.TLP.domain.Subcategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubcategoryRepository extends JpaRepository<Subcategory, Integer> {
+public interface SubcategoryRepository extends JpaRepository<Subcategory, Long> {
 }
